Lite-HRNet/src/requirement.txt
absl-py==1.0.0
addict==2.4.0
aiohttp @ file:///tmp/build/80754af9/aiohttp_1646806366512/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout @ file:///tmp/build/80754af9/async-timeout_1637851218186/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
blinker==1.4
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.5.18.1
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chumpy==0.70
click @ file:///tmp/build/80754af9/click_1646038465422/work
contextlib2==21.6.0
cryptography @ file:///tmp/build/80754af9/cryptography_1652083738073/work
cycler==0.11.0
Cython==0.29.30
debugpy @ file:///tmp/build/80754af9/debugpy_1637091796427/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
easydict==1.9
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926445639/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fonttools==4.33.3
frozenlist @ file:///tmp/build/80754af9/frozenlist_1637767111923/work
google-auth==2.7.0
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
grpcio==1.46.3
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562408398/work
iniconfig==1.1.1
ipykernel @ file:///tmp/build/80754af9/ipykernel_1646982648551/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///opt/conda/conda-bld/ipython_1651600145335/work
jedi @ file:///tmp/build/80754af9/jedi_1644315233700/work
joblib==1.1.0
json-tricks==3.15.5
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1650622202839/work
jupyter-core @ file:///opt/conda/conda-bld/jupyter_core_1651671229925/work
kiwisolver==1.4.2
Markdown @ file:///tmp/build/80754af9/markdown_1614363528767/work
matplotlib==3.5.2
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mindspore-gpu @ https://ms-release.obs.cn-north-4.myhuaweicloud.com/1.7.0/MindSpore/gpu/x86_64/cuda-11.1/mindspore_gpu-1.7.0-cp38-cp38-linux_x86_64.whl
mindvision==0.1.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
ml-collections==0.1.1
mmcv==1.5.3
mmpose==0.27.0
multidict @ file:///opt/conda/conda-bld/multidict_1640703752579/work
munkres==1.1.4
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1649847908682/work
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1643507977997/work
opencv-contrib-python-headless==4.5.5.64
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.2
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
pluggy==1.0.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.20.1
psutil @ file:///tmp/build/80754af9/psutil_1612298023621/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py==1.11.0
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycocotools==2.0.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1652398519695/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytest==7.1.2
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.1
PyYAML==6.0
pyzmq @ file:///tmp/build/80754af9/pyzmq_1638436375034/work
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-learn==1.0.2
scipy @ file:///tmp/build/80754af9/scipy_1641555001653/work
Shapely==1.8.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1651519171059/work/tensorboard-2.9.0-py3-none-any.whl
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.11.0
torch-tb-profiler==0.4.0
torchvision==0.12.0
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm==4.64.0
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1650639997961/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
xtcocotools==1.12
yacs==0.1.8
yapf==0.32.0
yarl @ file:///tmp/build/80754af9/yarl_1606939922162/work
zipp @ file:///opt/conda/conda-bld/zipp_1652341764480/work
